added rating bar and made the use of it

+++ app/src/main/java/fi/metropolia/retrofitparliamentmember/model/Review.kt
@@ -18,5 +18,5 @@ data class Review(
     val id: Int,
     val personId: Int,
     val comment: String,
-    val rating: Int
+    val rating: Float
 )
